mood ttl as int, show mood text

+++ rule-mood/rule-mood-web/src/main/java/com/liferay/content/targeting/rule/mood/display/context/MoodRuleDisplayContext.java
@@ -52,7 +52,7 @@ public class MoodRuleDisplayContext extends BaseRuleDisplayContext {
 		return _moodThreshold;
 	}
 
-	public double getMoodTtl() {
+	public int getMoodTtl() {
 		if (Validator.isNull(_moodTtl)) {
 			_moodTtl = GetterUtil.getInteger(context.get("moodTtl"));
 		}
